fix: Prevent tests to break when.
Not sure what's the best way of handling this. The tests using the in-memory database cannot be run together with the ones using the actual mongodb instance since something somehow breaks, but I don't understand exactly why. Until finding a better solution, tests should be grouped in modules and invoked separately, then all the coverage files combined (as before)